Mise a jour souscription

+++ AssoFlex/Models/ArticlePanier.cs
@@ -6,7 +6,7 @@ namespace AssoFlex.Models
         public int UtilisateurId { get; set; }
         public int ProduitId { get; set; }
         public string ProduitNom { get; set; }
-        public decimal MontantUnitaire { get; set; }
+        public double MontantUnitaire { get; set; }
         public int Quantite { get; set; }
         public string TypeDeCommande {get; set;}
         public int PanierId { get; set; }
